Move subcommands to own namespace for better organisation
